// src/shared/types/node.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GraphNodeSchema {
  /**
   * The node's ID, also used as node in an edges source/target etc.
   */
  key: string;
  /**
   * The node's attributes.
   */
  attributes: {
    type?: string;
    /**
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^(?!type$).*".
     */
    [k: string]: number | string | boolean | GraphInstantiatedAttributeSchema;
  };
}
export interface GraphInstantiatedAttributeSchema {
  type: string;
  args: {
    [k: string]: unknown;
  };
}